Refactor Utils/CreepUtils
