+++ source/js/buyers-guide/search.js
@@ -4,8 +4,8 @@
 
 import { SearchFilter } from "./search/search-filter.js";
 import { PNIToggle } from "./search/pni-toggle.js";
-import { PNISortCreepiness } from "./search/sort-creepiness.js";
+import { PNISortDropdown } from "./search/pni-sort-dropdown.js";
 
 const searchFilter = new SearchFilter();
 new PNIToggle(searchFilter);
-new PNISortCreepiness(searchFilter);
+new PNISortDropdown(searchFilter);
